Moving configuartion files into correct directory
Maven expect configuration to be located in src/main/resources and not src/resources. This is also why we were
having problems with the log4j configuation not being copied over. I also changed the java classes to look
for properties files in the classpath rather than a single specific directory. Also reconfigured the Maven
Flyway plugin to look for its configuration file in this global directory, so now all configuration files live
in the same place :)